feat: /auth 경로에 로그인 테스트 로직 구축
