+++ snapsplit/src/app/(no-gnb)/auth/page.tsx
@@ -6,7 +6,7 @@ export interface AuthProps {
 
 export default function Auth({}: AuthProps) {
   const REST_API_KEY = '5099e273f6583110b2cd3053783eddb4';
-  const REDIRECT_URI = 'http://localhost:3000/api/auth/callback/kakao';
+  const REDIRECT_URI = 'http://localhost:3000/oauth/callback/kakao';
 
   const kakaoLogin = () => {
     window.location.href = `https://kauth.kakao.com/oauth/authorize?client_id=${REST_API_KEY}&redirect_uri=${REDIRECT_URI}&response_type=code`;
